requirements.txt
slackclient==2.9.3
slackeventsapi==2.2.1
Flask==1.1.2
requests==2.25.1
google-api-python-client==2.0.2
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.3
pandas==1.2.3
penpyxl==3.0.7
xlsxwriter==1.3.7